refactor(vision): rename tool image resolver
